Create todo list project select and cacheDom
I added a select element to the todo list modal that lets a user select
which project to associate their list with. There is just a default
project until more are created. I also created a cache for dom elements
inside the bindEvents file so that I am not using query selector as
often for elements that are already available.