Refactor createDashboardElements into render fns
I decided to change the createDashboardElements function and break it
out into two render functions, renderBody and renderViewAllTodos. I
wanted to have the main elements like the content and sidebar be in
their own function and the view all todos to be on its own so that when
switching between on project and view all todos we are not rerendering
the base elements again.