Create functions show todo list and create item
The renderTodoList function will create the view for when a user clicks
on a todolist name in a project. For now it will just list out each item
and its fields. The createNewTodoItem makes a new todo item from the
fields on the new todo item modal and adds it to the todo list. Also
added some elements to cacheDom in bindEvents for new todo items.