Add theme colors and style most of solo list view
I created some variables for theme colors and changed most of the
existing colors to use the new colors. The majority of the single todo
list view has been completed, besides the modal for adding a new todo
item. Also the checkbox for completing a todo item might need a color
change.