Add collapsible todo items to todolist view
In the single todo list view, each todo item has an icon to click for
opening and closing the todo item revealing more of its data. I also
changed some styling for the view all todo lists to better handle list
sizes and show todo priority and if it is complete or not.