Bold project name in sidebar when selected
I made it so that when a project is clicked on the name of the project
in the sidebar will be bolded. When another project is clicked the
selected class is removed and added that other project, if view all
todos is clicked the selected class is just removed. Also added priority
styling to todo item names in the todo list view.