Create function to delete a project
Added icon next to project names to delete a project, but not Default
which will always be there. The event listener handles the views when
a project gets deleted and that project is being shown on the screen.
Also the view all todos view gets re-rendered to not show any todo lists
that were lost when the project was deleted.