Add localStorage functionality to save projects
I added a function to check if localStorage is available and then the
projects in Dashboard will be saved anytime you update them. If the
localStorage item is not found we create a new default project and then
save it in localStorage if available. This applies to all todo lists and
todo items inside the projects too.